% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a9a}
\alias{a9a}
\title{a9a}
\format{A dataset of 32,561 observations and 123 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
a9a
}
\description{
a9a
}
\keyword{datasets}
