% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covtype}
\alias{covtype}
\title{covtype}
\format{A dataset of 581,012 observations and 54 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
covtype
}
\description{
covtype
}
\keyword{datasets}
