% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_medium}
\alias{data_medium}
\title{Benchmarks for medium task}
\format{A data frame with 800 rows and 5 variables:
\describe{
\item{rows}{number of rows in data}
\item{cols}{number of columns in data}
\item{time}{time in nanoseconds}
\item{loss}{value of objective function}
\item{package}{package}
}}
\usage{
data_medium
}
\description{
A benchmark of L2-regularized logistic regression in the
\pkg{bigoptim} and \pkg{glmnet} packages based on subsets of the
\link[bigoptim:covtype.libsvm]{bigoptim::covtype.libsvm} dataset.
}
\keyword{datasets}
