% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ijcnn1}
\alias{ijcnn1}
\title{ijcnn1}
\format{A dataset of 49,990 observations and 22 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
ijcnn1
}
\description{
ijcnn1
}
\keyword{datasets}
