% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushrooms}
\alias{mushrooms}
\title{mushrooms}
\format{A dataset of 8,124 observations and 112 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
mushrooms
}
\description{
mushrooms
}
\keyword{datasets}
