% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective_r}
\alias{objective_r}
\title{R-based objective function for L1- or L2-regularized logistic regression}
\usage{
objective_r(beta0, beta, x, y, lambda, alpha = 0)
}
\arguments{
\item{beta0}{intercept}

\item{beta}{a feature matrix}

\item{x}{observations}

\item{y}{response}

\item{lambda}{penalty}

\item{alpha}{elasticnet mixing parameter}
}
\value{
Objective function value.
}
\description{
R-based objective function for L1- or L2-regularized logistic regression
}
\examples{
if (requireNamespace("glmnet")) {
  x <- matrix(rnorm(100*20), 100, 20)
  y <- sample(1:2, 100, replace = TRUE)
  fit <- glmnet::cv.glmnet(x, y, family = "binomial")
  lambda <- fit$lambda.1se
  beta0 <- coef(fit, lambda)[1]
  beta <- coef(fit, lambda)[-1]
  objective_r(beta0, beta, t(x), y, lambda, alpha = 1)
}
}
