% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phishing}
\alias{phishing}
\title{phishing}
\format{A dataset of 11,055 observations and 68 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
phishing
}
\description{
phishing
}
\keyword{datasets}
