% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skin_nonskin}
\alias{skin_nonskin}
\title{skin_nonskin}
\format{A dataset of 245,057 observations and 3 features.}
\source{
https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html
}
\usage{
skin_nonskin
}
\description{
skin_nonskin
}
\keyword{datasets}
