% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a9a}
\alias{a9a}
\title{a9a}
\format{A data frame with 800 rows and 5 variables:
\describe{
\item{rows}{number of rows in data}
\item{cols}{number of columns in data}
\item{time}{time in nanoseconds}
\item{loss}{value of objective function}
\item{package}{package}
}}
\usage{
a9a
}
\description{
a9a
}
\keyword{datasets}
