% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mediumhard}
\alias{data_mediumhard}
\title{Benchmarks for medium-hard task}
\format{A data frame with 800 rows and 5 variables:
\describe{
\item{dataset}{dataset}
\item{package}{package, glmnet or saga}
\item{time}{runtime in seconds}
\item{loss}{value of objective function}
}}
\usage{
data_mediumhard
}
\description{
A benchmark of SAGA L1-regularized logistic regression in the
\pkg{scikit-learn} python package and the python version of
\pkg{glmnet} based
on subsets of the \link{covtype}, \link{ijcnn1}, \link{a9a}, \link{mushrooms},
\link{skin_noskin}, and \link{phishing} datasets.
}
\keyword{datasets}
